schedstats: revert #6975, use 32Bit again

+++ core/include/sched.h
@@ -179,7 +179,7 @@ NORETURN void sched_task_exit(void);
  *  Scheduler statistics
  */
 typedef struct {
-    uint64_t laststart;      /**< Time stamp of the last time this thread was
+    uint32_t laststart;      /**< Time stamp of the last time this thread was
                                   scheduled to run */
     unsigned int schedules;  /**< How often the thread was scheduled to run */
     uint64_t runtime_ticks;  /**< The total runtime of this thread in ticks */
